print_agent/requirements.txt
websockets>=15.0
python-escpos>=3.1
pyusb>=1.3.1
pillow>=12.0
